+++ Soomla/Assets/Soomla/Editor/android/ManifestTools.cs
@@ -43,7 +43,6 @@ namespace UnityEditor.SoomlaEditor
 			
 			string ns = manNode.GetNamespaceOfPrefix("android");
 
-			findOrPrependElement("uses-permission", "name", ns, "com.android.vending.BILLING", manNode, doc);
 			findOrPrependElement("uses-permission", "name", ns, "android.permission.INTERNET", manNode, doc);
 
 			ns = applicationNode.GetNamespaceOfPrefix("android");
@@ -51,20 +50,86 @@ namespace UnityEditor.SoomlaEditor
 			XmlElement applicationElement = FindChildElement(manNode, "application");
 			applicationElement.SetAttribute("name", ns, "com.soomla.store.SoomlaApp");
 
-			findOrAppendIabActivity(ns, applicationNode, doc);
+			if (SoomSettings.GPlayBP) {
+				findOrPrependElement("uses-permission", "name", ns, "com.android.vending.BILLING", manNode, doc);
+				findOrAppendIabActivity(ns, applicationNode, doc);
+				findOrAppendMetaDataTag(ns, applicationNode, "google.GooglePlayIabService", doc);
+			} else {
+				findAndRemoveElement("uses-permission", "name", ns, "com.android.vending.BILLING", manNode, doc);
+				findAndRemoveIabActivity(ns, applicationNode, doc);
+			}
+
+			if (SoomSettings.AmazonBP) {
+				findOrAppendAmazonRR(ns, applicationNode, doc);
+				findOrAppendMetaDataTag(ns, applicationNode, "amazon.AmazonIabService", doc);
+			} else {
+				findAndRemoveAmazonRR(ns, applicationNode, doc);
+			}
 
 			doc.Save(fullPath);
 		}
 
+		private static void findOrAppendAmazonRR(string ns, XmlNode applicationNode, XmlDocument doc) {
+			XmlElement e = FindElementForNameWithNamespace("receiver", "name", ns, "com.amazon.inapp.purchasing.ResponseReceiver", applicationNode);
+			if (e == null)
+			{
+				e = doc.CreateElement("receiver");
+				e.SetAttribute("name", ns, "com.amazon.inapp.purchasing.ResponseReceiver");
+				e.InnerText = "\n    ";
+
+				XmlElement intentE = doc.CreateElement("intent-filter");
+				intentE.InnerText = "\n    ";
+
+				XmlElement actionE = doc.CreateElement("action");
+				actionE.SetAttribute("name", ns, "com.amazon.inapp.purchasing.NOTIFY");
+				actionE.SetAttribute("permission", ns, "com.amazon.inapp.purchasing.Permission.NOTIFY");
+				actionE.InnerText = "\n    ";
+
+				intentE.AppendChild(actionE);
+				e.AppendChild(intentE);
+				applicationNode.AppendChild(e);
+			}
+		}
+
+		private static void findAndRemoveAmazonRR(string ns, XmlNode applicationNode, XmlDocument doc) {
+			XmlElement e = FindElementForNameWithNamespace("receiver", "name", ns, "com.amazon.inapp.purchasing.ResponseReceiver", applicationNode);
+			if (e != null)
+			{
+				applicationNode.RemoveChild(e);
+			}
+		}
+
 		private static void findOrAppendIabActivity(string ns, XmlNode applicationNode, XmlDocument doc) {
-			XmlElement e = FindElementForNameWithNamespace("activity", "name", ns, "com.soomla.store.StoreController$IabActivity", applicationNode);
+			XmlElement e = FindElementForNameWithNamespace("activity", "name", ns, "com.soomla.store.billing.google.GooglePlayIabService$IabActivity", applicationNode);
 			if (e == null)
 			{
 				e = doc.CreateElement("activity");
-				e.SetAttribute("name", ns, "com.soomla.store.StoreController$IabActivity");
+				e.SetAttribute("name", ns, "com.soomla.store.billing.google.GooglePlayIabService$IabActivity");
 				e.SetAttribute("theme", ns, "@android:style/Theme.Translucent.NoTitleBar.Fullscreen");
 				e.InnerText = "\n    ";
 				applicationNode.AppendChild(e);
+			}
+		}
+
+		private static void findAndRemoveIabActivity(string ns, XmlNode applicationNode, XmlDocument doc) {
+			XmlElement e = FindElementForNameWithNamespace("activity", "name", ns, "com.soomla.store.billing.google.GooglePlayIabService$IabActivity", applicationNode);
+			if (e != null)
+			{
+				applicationNode.RemoveChild(e);
+			}
+		}
+
+		private static void findOrAppendMetaDataTag(string ns, XmlNode applicationNode, string bpProvider, XmlDocument doc) {
+			XmlElement e = FindElementForNameWithNamespace("meta-data", "name", ns, "billing.service", applicationNode);
+			if (e == null)
+			{
+				e = doc.CreateElement("meta-data");
+				e.SetAttribute("name", ns, "billing.service");
+				e.SetAttribute("value", ns, bpProvider);
+				e.InnerText = "\n    ";
+				applicationNode.AppendChild(e);
+			} else {
+				e.SetAttribute("value", ns, bpProvider);
 			}
 		}
 
@@ -75,6 +140,14 @@ namespace UnityEditor.SoomlaEditor
 				e = doc.CreateElement(name);
 				e.SetAttribute(androidName, ns, value);
 				parent.PrependChild(e);
+			}
+		}
+
+		private static void findAndRemoveElement(string name, string androidName, string ns, string value, XmlNode parent, XmlDocument doc) {
+			XmlElement e = FindElementForNameWithNamespace(name, androidName, ns, value, parent);
+			if (e != null)
+			{
+				parent.RemoveChild(e);
 			}
 		}